Added basic support for drag and drop.
The main functionality is implemented in AbstractSWTBot#dragAndDrop(target), which makes it available for virtually
any widget supported by SWTBot. The only requirement is that absoluteLocation() must be implemented for both source
and target widgets. The code bases on the example outlined here: http://osdir.com/ml/java.abbot.user/2006-03/msg00043.html

It seems like we have to use native events as e.g. DoDragDrop is used by SWT internally under windows:

    http://msdn.microsoft.com/en-us/library/ms678486%28VS.85%29.aspx

The native method runs a loop waiting for more "dnd events".

For SWTBotList, there is SWTBotList#dragAndDrop(item, target), allowing to drag a specific item instead of
the one which happens to be located at the drag start position. This special treatment was necessary as a List
only contains an array of Strings, there it no such thing as a ListItem Control as a list's child (as opposed
to a Tree's TreeItems).

For Trees, you will simply use the respective TreeItem as "drag source".

Check org.eclipse.swtbot.swt.finder.widgets.SWTBotDragAndDropTest for a few examples.

Tested under windows only.