Allow for indication of a drop location
Added the following public methods to AbstractSWTBot:
          void  dragAndDrop(Point)
   static Point before(AbstractSWTBot<T>)
   static Point on(AbstractSWTBot<T>)
   static Point after(AbstractSWTBot<T>)

This allows users to write things like:

  source.dragAndDrop(after(target));
  source.dragAndDrop(on(target));  // equal to source.dragAndDrop(target)
  source.dragAndDrop(before(target));

While "on" is the default behavior if only the target widget is specified.

This also makes dropping to an arbitrary location publicly available
(since anyone can create a new Point() and hand it over to dragAndDrop).
Which might be a problem but might also be helpful in some situations.

Another problem is the decision about inserting items before, after or
below other ones. This is totally up to the DropTarget implementation.
For this test I used a DND Snippet (http://dev.eclipse.org/viewcvs/index.cgi/org.eclipse.swt.snippets/src/org/eclipse/swt/snippets/Snippet91.java?view=co).
This implementation splits the "target item"'s bounds into three areas.
Items dropped on the upper third will be added before the selected item.
Items dropped on the middle third will be added as child to the selected item.
Items dropped on the lower third will be added after the selected item.
Based on this, before() returns a location on the upper border of the target,
on() its center point and after() a location on the lower border of the target.
The location specifiers work well with this implementation under windows.

The default SWT Example for DND does not consider this at all. Items are
always added as children. They are only appended if dropped somewhere after
the last item in the tree.